Move DomainManager into Internal namespace and folder
