+++ src/TestEngine/testcentric.engine.core/Runners/TestDomainRunner.cs
@@ -5,7 +5,7 @@
 
 #if !NETSTANDARD1_6 && !NETSTANDARD2_0
 using NUnit.Engine;
-using TestCentric.Engine.Services;
+using TestCentric.Engine.Internal;
 
 namespace TestCentric.Engine.Runners
 {
